fix loading a conversation message
this was failing because it was doing
_.pluck(message.participants, 'name')
but a bunch of message.participants were undefined

I just saw this because it was not working for me
in my own canvas.
The best way I can think of to show you how this
was broken is this screenshot:
http://cl.ly/image/3E061r2F2B3W

test plan:
open new conversations inbox.
click a message,
it should load

Change-Id: I7273d92d8bd206da0b2c56fabf1951bdc46f3095
Reviewed-on: https://gerrit.instructure.com/35214
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Braden Anderson <banderson@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>