make wiki sidebar images tab accessible
fixes #CNVS-12824

Images in the wikisidebar were not accessible, they now are. We also
removed a few links at the bottom of the page that were not needed.

Test Plan
----------------
As a user
Given I'm editing an assignment
And I tab to the wiki sidebar's "image" tab
Then I should be able to tab to images
And I should be able to press enter to insert an image into the page

Test Plan
----------------
As a user
Given I'm editing an assignment
Then I should not see the "New Discussion" link
And I should not see the "New Quiz" link

Change-Id: Icd5f1fc66c83000a05a00c9a54dc9ea948cfa639
Reviewed-on: https://gerrit.instructure.com/36102
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>