only parse integration_data when string
When assignments that contain integration_data are saved, they
currently fail on save if integration_data is not part of the
save--i.e. since integration_data is already a parsed Hash, and
our code tries to parse it as JSON, the JSON parser raises an
exception.

This prevents JSON parsing when integration_data is not a string.

Fixes SIS-375

Change-Id: I305479c5a9999c0b7a452099d598467f70079e80
Reviewed-on: https://gerrit.instructure.com/38080
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Liz Abinante <labinante@instructure.com>
Reviewed-by: Tyler Pickett <tpickett+gerrit@instructure.com>
Product-Review: Duane Johnson <duane@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>