get c-common.scss parseable by node-sass
This is the hardest one, it was not too bad though.
Things I found you need to do:

1. You can't have a "//" comment on the same line in .sass syntax
so the following line:

.foo //a comment about foo

would need to be:
//a comment about #foo
.foo

2. in .sass, you have to use double quotes for @import
so:      @import 'foo'
becomes: @import "foo"

things you do NOT have to do:

1. don't have to add the extension or the underscore
in front of filename for @import.
e.g, DON'T have to convert:
@import "foo"
to:
@import "_foo.scss"

2: don't have to convert the leading colon syntax
   of super old sass to colon between syntax
e.g:
.foo
  :width 100%
is just fine

(Hallelujah on not having to do those 2, it
makes this a lot smaller of a change)

test plan:
`npm install node-sass` then
within the /app/stylesheets directory,
run this command and see that it works:

../../node_modules/.bin/node-sass --include-path variants/new_styles_high_contrast base/c-common.scss

Change-Id: I73e382560e3dcbf55c32e97bb245b5ec262d4cb0
Reviewed-on: https://gerrit.instructure.com/38414
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>