remove vendor prefixes from flexboxgrid.scss
autoprefixer takes care of these for us. so it's
code we don't need, less to maintain.

But if you want to just leave it as-is so it's
easier to diff against the original that's fine
too. I'll abandon this.

FYI, the regex I used was .*-(ms|webkit|moz).*\n

test plan:
the outputted css should be exactly the same as
what it used to be.

Change-Id: I3178dd220f5f1bc67c579298ad730c6b1a8b9a95
Reviewed-on: https://gerrit.instructure.com/40622
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Chris Hart <chart@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>