log out specific cas ticket
fixes: CNVS-15843

Changes how we associate a CAS ticket with the Canvas session.  When we
recieve a logout request from the CAS server we mark the ticket as
expired and each time the user performes a request we check their
session cas ticket to the ticket we stored to see if its expired or not.

Test Plan:

 - Setup CAS Server
 - Log in to Canvas with a user
 - Log in to Canvas with the same user in a different browser.
 - Have the CAS server send a SLO request for one of the tickets.
 - First session should be logged out on the next request.
 - Second session should not be logged out.

For additional details see:

https://github.com/instructure/canvas-lms/pull/516#issuecomment-57340629

Change-Id: I189f70e58db6444b1747a686a6a50f46c53f3697
Reviewed-on: https://gerrit.instructure.com/42312
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>