add force_token_reuse to developer key
API consumers should generally  try to reuse the api tokens that
they generate. This property makes it so that new tokens
won't be issued to a developer key for a user if that user
has a token that is still valid for that key.

Test Plan:
 * for both a trusted and an unstrusted developer key:
 * without force_token_reuse set on the key:
 * do the token request flow twice, you should get a new token each time
 * with force_token_reuse set:
 * do the token request flow twice, the second time access_token should be blank

closes PLAT-767

Change-Id: I94e7dfbe24ce534cca0ecba940c1e0b3af15b0cd
Reviewed-on: https://gerrit.instructure.com/44393
Reviewed-by: Dave Donahue <ddonahue@instructure.com>
Tested-by: Jenkins
QA-Review: August Thornton <august@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>