Adds an iterator to duplicate file names
When a file is uploaded via the WikiSidebar,
it can be uploaded multiple times with the same
name.  This commit adds an iterator to uploaded
files so that if the file has already been uploaded
or a file with the same name already exists,
an iterator is added to the newly uploaded file
to prevent duplicates (e.g. file.jpg, file-1.jpg, file-2.jpg).

fixes CNVS-17361

Test Plan:
* Enable "Better File Browsing" feature flag
* Go to any course's Pages section (as admin/teacher)
* Create a new page or edit an existing one
* On the create/edit page, select the "Files"
  tab from the WikiSidebar on the right
* Click "+ Upload a new file"
* Select some file, then upload it
* Upload the same file again, multiple times
* Go to that course's Files page
* In the course files (or wherever you chose to
  upload the file), the file you uploaded should be
  there, along with additional files with the same name
  followed by an iterator (e.g. yourfile.png, yourfile-1.png,
  yourfile-2.png, etc)

Change-Id: Iaa1307305659bf3b1b38991f0981ac6e48423209
Reviewed-on: https://gerrit.instructure.com/45550
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Dan Minkevitch <dan@instructure.com>